% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designers.R
\name{get_designer}
\alias{get_designer}
\title{Get details for a designer}
\usage{
get_designer(id)
}
\arguments{
\item{id}{designer id}
}
\value{
tibble with name, notes, count of patterns, etc.
}
\description{
This function retrieves details for a designer using their id.
}
\examples{
\dontrun{get_designer(id = 1)}

}
