% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get_pattern_categories}
\alias{get_pattern_categories}
\title{Get pattern categories}
\usage{
get_pattern_categories()
}
\value{
nested tibble containing levels of categories and sub-categories for patterns.
}
\description{
This function retrieves pattern categories.
}
\examples{
\dontrun{ get_pattern_categories() }

}
