% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get_pattern_comments}
\alias{get_pattern_comments}
\title{Get comments for a pattern}
\usage{
get_pattern_comments(id, page = NULL, page_size = NULL, sort = NULL)
}
\arguments{
\item{id}{a `pattern_id`}

\item{page}{result page to retrieve; defaults to first page}

\item{page_size}{number of results to retrieve; defaults to 25, max is 100}

\item{sort}{sort order; options are `time`, `helpful`
(and `time_`, `helpful_` for descending sort)}
}
\value{
tibble
}
\description{
This function retrieves comments for a pattern using the pattern id.
}
\examples{
\dontrun{get_pattern_comments(id = 600, page_size = 10)}

}
