% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get_patterns}
\alias{get_patterns}
\title{Get pattern details for multiple patterns}
\usage{
get_patterns(ids)
}
\arguments{
\item{ids}{one or more `pattern_id`s}
}
\value{
tibble containing pattern details, like yardage, gauge, count of comments, ratings, etc.
}
\description{
This function retrieves details for one or multiple patterns using the pattern id.
}
\examples{
get_patterns(ids = c(600, 601))

}
