% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shops.R
\name{get_shop}
\alias{get_shop}
\title{Get details for a shop}
\usage{
get_shop(id)
}
\arguments{
\item{id}{shop id}
}
\value{
tibble containing shop info including address, contact info, and other descriptive variables.
}
\description{
This function retrieves details for a shop using its id.
}
\examples{
\dontrun{get_shop(id = 1781)}

}
