% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{get_yarns}
\alias{get_yarns}
\title{Get details for multiple yarns}
\usage{
get_yarns(ids)
}
\arguments{
\item{ids}{one or more `yarn_id`s}
}
\value{
tibble
}
\description{
This function retrieves details for one or multiple yarns using the yarn id.
}
