% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravelry_auth.R
\name{ravelry_auth}
\alias{ravelry_auth}
\title{Get or set basic authentication credentials}
\usage{
ravelry_auth(key = "username", overwrite = FALSE)
}
\arguments{
\item{key}{the variable to set - either username or password}

\item{overwrite}{overwrite the existing variable in the current environment?}
}
\value{
atomic character vector containing the username or password credential
}
\description{
The API wrapper functions in this package rely on basic authentication via username
and password residing in the environment variables \code{RAVELRY_USERNAME} and
\code{RAVELRY_PASSWORD}. The easiest way to accomplish this is to set it in
the `\code{.Renviron}` file in your home directory. If you do not have these
credentials, you can create them at https://www.ravelry.com/pro/developer.
}
