% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ravelry_get.R
\name{ravelry_get}
\alias{ravelry_get}
\title{Basic API call}
\usage{
ravelry_get(path, query = NULL)
}
\arguments{
\item{path}{the API method to call}

\item{query}{call parameters in the form of a list()}
}
\value{
API object
}
\description{
This function is the base GET call to the Ravelry API.
}
