% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{search_patterns}
\alias{search_patterns}
\title{Search for patterns}
\usage{
search_patterns(query, page = NULL, page_size = NULL)
}
\arguments{
\item{query}{search string}

\item{page}{result page to retrieve; defaults to first page}

\item{page_size}{number of results to retrieve; defaults to 100}
}
\value{
tibble containing basic pattern details, some nested
}
\description{
This function retrieves a list of pattern ids and basic details based on search terms.
}
\examples{
search_patterns(query = 'cowl')

}
