% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{search_yarn}
\alias{search_yarn}
\title{Search for yarn}
\usage{
search_yarn(query, page = NULL, page_size = NULL, sort = NULL)
}
\arguments{
\item{query}{search string}

\item{page}{result page to retrieve; defaults to first page}

\item{page_size}{number of results to retrieve; defaults to 100}

\item{sort}{options: `best`, `rating`, `projects`}
}
\value{
Tibble containing basic pattern details, some nested
}
\description{
This function retrieves a list of basic yarn details based on search terms.
}
