% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get_patterns}
\alias{get_patterns}
\title{Get pattern details for multiple patterns}
\usage{
get_patterns(ids)
}
\arguments{
\item{ids}{one or more `pattern_id`s}
}
\value{
API pattern object
}
\description{
This function retrieves details for one or multiple patterns using the pattern id.
}
