% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarns.R
\name{get_yarns}
\alias{get_yarns}
\title{Get details for multiple yarns}
\usage{
get_yarns(ids)
}
\arguments{
\item{ids}{one or more `yarn_id`s}
}
\value{
tibble containing yarn details like company, gauge, grams, texture, needle sizes, ratings, etc.
}
\description{
This function retrieves details for one or multiple yarns using the yarn id.
}
\examples{
\dontrun{get_yarns(ids = c(66124,54110))}

}
