% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{search_patterns}
\alias{search_patterns}
\title{Search for patterns}
\usage{
search_patterns(query = NULL, page = NULL, page_size = NULL, ...)
}
\arguments{
\item{query}{search string}

\item{page}{result page to retrieve; defaults to first page}

\item{page_size}{number of results to retrieve; defaults to 100}

\item{...}{pass any other filter parameters available via \url{https://www.ravelry.com/patterns/search}
(ex. craft, availability, photo, pa, fit, weight, etc.)}
}
\value{
tibble containing basic pattern details, some nested
}
\description{
This function retrieves a list of pattern ids and basic details based on search terms.
}
\examples{
\dontrun{search_patterns(query = 'cowl')}

# with site search parameters
\dontrun{search_patterns(query = 'hat', page_size = 10, availability = 'free', fit = 'baby')}

}
