% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarns.R
\name{search_yarn}
\alias{search_yarn}
\title{Search for yarn}
\usage{
search_yarn(query = NULL, page = NULL, page_size = NULL,
  sort = NULL, ...)
}
\arguments{
\item{query}{search string}

\item{page}{result page to retrieve; defaults to first page}

\item{page_size}{number of results to retrieve; defaults to 100}

\item{sort}{options: `best`, `rating`, `projects`}

\item{...}{pass any other filter parameters available via \url{https://www.ravelry.com/yarns/search}
(ex. weight, needles, photo, fiberc, ya, ratings, origin)}
}
\value{
tibble containing basic pattern details, some nested
}
\description{
This function retrieves a list of basic yarn details based on search terms.
}
\examples{
\dontrun{search_yarn(query = 'cascade', page_size = 10)}

# using site search parameters
\dontrun{search_yarn(query = 'cascade', sort = 'best', weight = 'sport', needles = '3.75mm')}

}
