% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarns.R
\name{search_yarn_companies}
\alias{search_yarn_companies}
\title{Search yarn companies}
\usage{
search_yarn_companies(query = NULL, page = NULL, page_size = NULL,
  sort = NULL, ...)
}
\arguments{
\item{query}{search string}

\item{page}{result page to retrieve; defaults to first page}

\item{page_size}{number of results to retrieve; defaults to 48}

\item{sort}{options: `best`}

\item{...}{pass any other filter parameters available via \url{https://www.ravelry.com/yarns/brands/search}}
}
\value{
tibble containing company details
}
\description{
This function retrieves yarn companies based on search terms.
}
\examples{
\dontrun{search_yarn_companies(query = 'lion')}

}
